reorganized pdf url lookup code under /lib/forms/submission_statuses
